% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_country_name.R
\name{align_country_name}
\alias{align_country_name}
\title{Function to order country name}
\usage{
align_country_name(x)
}
\arguments{
\item{x}{Character vector with country names}
}
\description{
Function to order country name
}
\examples{
align_country_name(c("US", "UK", "Россия")) ## returns c("USA", "United Kingdom", "Russia")


}
