% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electroprice2019}
\alias{electroprice2019}
\title{Electroenergy price by countries}
\format{A data frame with 231 rows and 8 variables:
\describe{
\item{counry}{Country name}
\item{currency}{Currency name aligned with \code{\link{align_country_name}}}
\item{household_kwh}{Electricity prices for households}
\item{business_kwh}{Electricity prices for business}
\item{year}{Year by information source}
\item{descr_html}{Description by www.globalpetrolprices.com}
\item{data_source_read_date}{Scraping date}
...
}}
\source{
\url{https://www.globalpetrolprices.com/electricity_prices/}
}
\usage{
electroprice2019
}
\description{
A dataset containing the prices of electroenergy by counry in US dollars
and counry currency. You can srap this dataframe by \code{\link{scrap_globalpetrolprices_com}}
}
\keyword{datasets}
