% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalpetrolprices_country_elecroenergy_price.R
\name{scrap_globalpetrolprices_com}
\alias{scrap_globalpetrolprices_com}
\title{Scraping data from www.globalpetrolprices.com}
\usage{
scrap_globalpetrolprices_com()
}
\description{
We use https://www.globalpetrolprices.com/electricity_prices/ url to download
electroenegy prices by counry
}
