% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_wiki_counry_pop.R
\name{scrap_wiki_counry_pop}
\alias{scrap_wiki_counry_pop}
\title{Scraping wikipedia country population data}
\usage{
scrap_wiki_counry_pop()
}
\description{
Sorce url
https://en.wikipedia.org/wiki/List_of_countries_and_dependencies_by_population
}
