% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_globalpetrolprices_com.R
\name{scrap_globalpetrolprices_com}
\alias{scrap_globalpetrolprices_com}
\title{Scraping data from www.globalpetrolprices.com}
\usage{
scrap_globalpetrolprices_com()
}
\description{
We use \href{https://www.globalpetrolprices.com/electricity_prices/}{www.globalpetrolprices.com}
to download electricity prices by country (\code\link{electroprice2019})
}
